:passport_control: Make sure that keycloak mandates are synced correctly
